requirements.txt
tqdm
python-dotenv
numpy
together
transformers
pygame
sentencepiece
matplotlib
pandas